Hdf float64 (#505)
* Store configurations and DMC weights in double precision

h5py.create_dataset defaults to float32 when given a shape and no dtype, so
initialize_hdf was silently rounding walker coordinates to single precision on
the way to disk, as was the DMC weights dataset. Everything is computed in
float64, so a restart came back with coordinates perturbed by ~5e-7 bohr rather
than the ones that were written.

These datasets now declare the dtype of the array they store. A restart from a
new file now reproduces its walkers exactly.

Old files still load, and load_hdf warns when the stored type differs from the
computed one, since to_hdf writes into the existing dataset and would keep
storing float32 for the rest of that calculation without saying so.

For scale, on LiH/ccECP with 500 walkers the single precision round trip shifted
local energies by 2e-8 Ha on average and the mean energy by 2e-9 Ha, against a
statistical error of 1.5e-2 Ha


* Test that configurations far from the origin round trip exactly

Single precision is relative, so the rounding error the previous commit removes
grows linearly with the magnitude of the coordinates: measured on an H2 chain,
the largest walker displacement from a float32 round trip goes from 9.5e-7 bohr
at the origin to 3.0e-5 bohr with the system translated 1000 bohr out, and the
largest local energy shift from 4.2e-6 to 8.4e-4 Ha. That is the regime of a
large supercell, so the test pins it.


---------

+++ pyqmc/method/dmc.py
@@ -385,7 +385,7 @@ def dmc_file(hdf_file, data, attr, configs, weights):
                 hdftools.setup_hdf(hdf, data, attr)
                 configs.initialize_hdf(hdf)
             if "weights" not in hdf.keys():
-                hdf.create_dataset("weights", weights.shape)
+                hdf.create_dataset("weights", weights.shape, dtype=weights.dtype)
             hdftools.append_hdf(hdf, data)
             configs.to_hdf(hdf)
             hdf["weights"][:] = weights
